+++ Controllers/UsersController.cs
@@ -1,13 +1,14 @@
 ﻿using DotNetCardsServer.Auth;
 using DotNetCardsServer.Exceptions;
-using DotNetCardsServer.Models.MockData;
 using DotNetCardsServer.Models.Users;
 using DotNetCardsServer.Services.Users;
 using DotNetCardsServer.Utils;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using MongoDB.Driver;
 using System.ComponentModel.DataAnnotations;
 using System.Numerics;
+using System.Security.Claims;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
@@ -26,17 +27,26 @@ namespace DotNetCardsServer.Controllers
 
         // GET: api/<UsersController>
         [HttpGet]
+        [Authorize(Policy = "MustBeAdmin")]
         public async Task<IActionResult> Get()
-        {
-            Console.WriteLine(HttpContext.User.FindFirst("id").Value);
+        {    
             List<User> users =await _usersService.GetUsersAsync();
             return Ok(users);
         }
 
+
         // GET api/<UsersController>/5
+        [Authorize]
         [HttpGet("{id}")]
         public async Task<IActionResult> Get(string id)
         {
+            bool.TryParse(HttpContext.User.FindFirstValue("isAdmin"), out bool isAdmin);
+            bool isMyId = HttpContext.User.FindFirstValue("id") == id;
+            if(!isAdmin && !isMyId)
+            {
+                return Unauthorized("You can watch only your own profile");
+            }
+
             try
             {
             User? u = await _usersService.GetOneUserAsync(id);
@@ -76,6 +86,13 @@ namespace DotNetCardsServer.Controllers
         [HttpPut("{id}")]
         public async Task<IActionResult> Put(string id, [FromBody] User updatedUser)
         {
+            bool.TryParse(HttpContext.User.FindFirstValue("isAdmin"), out bool isAdmin);
+            bool isMyId = HttpContext.User.FindFirstValue("id") == id;
+            if (!isAdmin && !isMyId)
+            {
+                return Unauthorized("You can watch only your own profile");
+            }
+
             try
             {
             User newUser = await _usersService.EditUserAsync(id, updatedUser);
@@ -91,6 +108,13 @@ namespace DotNetCardsServer.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(string id)
         {
+            bool.TryParse(HttpContext.User.FindFirstValue("isAdmin"), out bool isAdmin);
+            bool isMyId = HttpContext.User.FindFirstValue("id") == id;
+            if (!isAdmin && !isMyId)
+            {
+                return Unauthorized("You can watch only your own profile");
+            }
+
             try
             {
                 await _usersService.DeleteUserAsync(id);
